# public/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "OpenAPI Labkita",
        "version": "1.0.0",
        "description": "API untuk blog, tools, jadwal sholat, waktu, profil, produk, layanan, Al-Qur'an, dan wilayah Indonesia."
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [],
    "servers": [
        {
            "url": "https://api-blog.labkita.my.id",
            "description": "production"
        },
        {
            "url": "http://localhost:8080",
            "description": "local"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "summary": "Home",
                "tags": ["Health Check"],
                "responses": {
                    "200": {
                        "description": "Home page"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Login user",
                "tags": ["Auth"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "required": [
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil"
                    }
                }
            }
        },
        "/api/posts": {
            "get": {
                "summary": "Daftar semua post",
                "tags": ["Posts"],
                "parameters":[
                    {
                        "name": "sortDir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "desc"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar post"
                    }
                }
            },
            "post": {
                "summary": "Buat post baru (auth)",
                "tags": ["Posts"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "title"
                                    }
                                },
                                "required": [
                                    "title"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post berhasil dibuat"
                    }
                }
            }
        },
        "/api/admin/posts": {
            "get": {
                "summary": "Daftar semua post termasuk unpublish (auth)",
                "description": "Sama seperti GET /api/posts namun TIDAK memfilter published_at, sehingga post draft/unpublish ikut tampil. Memerlukan autentikasi.",
                "tags": ["Posts"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortDir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "desc"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "published_at"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar post termasuk unpublish"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/admin/posts/{postSlug}": {
            "get": {
                "summary": "Detail post termasuk unpublished (auth)",
                "description": "Sama seperti GET /api/posts/{postSlug} namun TIDAK memfilter published_at, sehingga post draft/unpublish ikut tampil. Memerlukan autentikasi.",
                "tags": ["Posts"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "postSlug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail post termasuk unpublished"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/posts/{postSlug}": {
            "get": {
                "summary": "Detail post",
                "tags": ["Posts"],
                "parameters": [
                    {
                        "name": "postSlug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail post"
                    }
                }
            },
            "put": {
                "summary": "Update post (auth)",
                "tags": ["Posts"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "postSlug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post berhasil diupdate"
                    }
                }
            },
            "delete": {
                "summary": "Hapus post (auth)",
                "tags": ["Posts"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "postSlug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post berhasil dihapus"
                    }
                }
            }
        },
        "/api/uploads/presign": {
            "post": {
                "summary": "Presign URL upload gambar ke Cloudflare R2 (auth)",
                "description": "Menghasilkan presigned PUT URL agar frontend bisa upload gambar langsung ke Cloudflare R2. Format yang didukung: jpg, jpeg, png, webp. Ukuran maksimal 1MB. URL kadaluarsa dalam 5 menit. Setelah upload, kirim nilai `image_path` sebagai `image_url` saat membuat/mengubah post.",
                "tags": ["Uploads"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "filename": {
                                        "type": "string",
                                        "description": "Nama file beserta ekstensi (jpg, jpeg, png, webp)",
                                        "example": "foto.jpg"
                                    },
                                    "size": {
                                        "type": "integer",
                                        "format": "int64",
                                        "description": "Ukuran file dalam byte. Maksimal 1048576 (1MB).",
                                        "maximum": 1048576,
                                        "example": 524288
                                    }
                                },
                                "required": [
                                    "filename",
                                    "size"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Presigned URL berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OK"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "upload_url": {
                                                    "type": "string",
                                                    "description": "Presigned URL untuk upload (PUT) ke R2",
                                                    "example": "https://<account_id>.r2.cloudflarestorage.com/webapp/blog/laravelblog/storage/abc123.jpg?X-Amz-Algorithm=..."
                                                },
                                                "method": {
                                                    "type": "string",
                                                    "example": "PUT"
                                                },
                                                "headers": {
                                                    "type": "object",
                                                    "description": "Header yang wajib dikirim saat upload (mis. Content-Type)",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    },
                                                    "example": {
                                                        "Content-Type": "image/jpeg"
                                                    }
                                                },
                                                "key": {
                                                    "type": "string",
                                                    "description": "Object key lengkap di R2 (termasuk prefix)",
                                                    "example": "blog/laravelblog/storage/abc123.jpg"
                                                },
                                                "image_path": {
                                                    "type": "string",
                                                    "description": "Nilai yang dikirim sebagai image_url saat create/update post (bagian setelah prefix)",
                                                    "example": "abc123.jpg"
                                                },
                                                "public_url": {
                                                    "type": "string",
                                                    "description": "URL publik gambar setelah berhasil diupload",
                                                    "example": "https://webassets.labkita.my.id/blog/laravelblog/storage/abc123.jpg"
                                                },
                                                "expires_in": {
                                                    "type": "integer",
                                                    "description": "Masa berlaku presigned URL dalam detik",
                                                    "example": 300
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Format tidak didukung atau ukuran melebihi 1MB"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/tools": {
            "get": {
                "summary": "Daftar tools",
                "tags": ["Tools"],
                "parameters":[
                    {
                        "name": "sortDir",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "desc"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "id"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "dev tools"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar tools"
                    }
                }
            },
            "post": {
                "summary": "Tambah tool (auth)",
                "tags": ["Tools"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Tool berhasil ditambah"
                    }
                }
            }
        },
        "/api/tools/{toolId}": {
            "put": {
                "summary": "Update tool (auth)",
                "tags": ["Tools"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "toolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tool berhasil diupdate"
                    }
                }
            },
            "delete": {
                "summary": "Hapus tool (auth)",
                "tags": ["Tools"],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "toolId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tool berhasil dihapus"
                    }
                }
            }
        },
        "/api/prayers": {
            "post": {
                "summary": "Jadwal sholat",
                "tags": ["Jadwal sholat"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "year": {
                                        "type": "number",
                                        "example": 2025
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "Asia/Jakarta"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "example": 6.2
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "example": 106.8
                                    }
                                },
                                "required": [
                                    "year"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jadwal sholat"
                    }
                }
            }
        },
        "/api/hijri/calendar": {
            "get": {
                "summary": "Kalender Hijriyah",
                "tags": ["Kalender Hijriyah"],
                "responses": {
                    "200": {
                        "description": "Kalender Hijriyah"
                    }
                }
            }
        },
        "/api/timezones": {
            "get": {
                "summary": "Daftar zona waktu",
                "tags": ["Zona waktu"],
                "responses": {
                    "200": {
                        "description": "Daftar zona waktu"
                    }
                }
            }
        },
        "/api/profile": {
            "get": {
                "summary": "Profil user",
                "tags": ["Profil"],
                "responses": {
                    "200": {
                        "description": "Profil user"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "summary": "Daftar produk",
                "tags": ["Produk"],
                "responses": {
                    "200": {
                        "description": "Daftar produk"
                    }
                }
            }
        },
        "/api/products/{productId}": {
            "get": {
                "summary": "Detail produk",
                "tags": ["Produk"],
                "parameters": [
                    {
                        "name": "productId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail produk"
                    }
                }
            }
        },
        "/api/services": {
            "get": {
                "summary": "Daftar layanan",
                "tags": ["Layanan"],
                "responses": {
                    "200": {
                        "description": "Daftar layanan"
                    }
                }
            }
        },
        "/api/surah": {
            "get": {
                "summary": "Daftar surah Al-Qur'an",
                "tags": ["Al-Qur'an"],
                "responses": {
                    "200": {
                        "description": "Daftar surah"
                    }
                }
            }
        },
        "/api/ayah": {
            "get": {
                "summary": "Detail ayat Al-Qur'an",
                "tags": ["Al-Qur'an"],
                "parameters": [
                    {
                        "name": "surahId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail ayat"
                    }
                }
            }
        },
        "/api/provinces": {
            "get": {
                "summary": "Daftar provinsi",
                "tags": ["Wilayah"],
                "responses": {
                    "200": {
                        "description": "Daftar provinsi"
                    }
                }
            }
        },
        "/api/cities": {
            "get": {
                "summary": "Daftar kota",
                "tags": ["Wilayah"],
                "parameters":[
                    {
                        "name": "provinceId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kota"
                    }
                }
            }
        },
        "/api/districts": {
            "get": {
                "summary": "Daftar kecamatan",
                "tags": ["Wilayah"],
                "parameters":[
                    {
                        "name": "cityId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "32.04"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kecamatan"
                    }
                }
            }
        },
        "/api/subdistricts": {
            "get": {
                "summary": "Daftar kelurahan",
                "tags": ["Wilayah"],
                "parameters":[
                    {
                        "name": "districtId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "32.04.37"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar kelurahan"
                    }
                }
            }
        }
    }
}